A L8Runtime can now actually be deallocated. Verified use of GetCurrentContext() in callbacks.
Because of the strong reference to the WrapperMap, and the WrapperMap containing a strong reference to the Runtime, plus the strong reference in the v8::Context, a Runtime would never deallocate. #1.
The use of GetCurrentContext() on the Isolate is correct, as in callbacks, the current context is always the context the callback came from. #5. (v8-users mailing list)